+++ zukit/src/scripts/components/blocks-index.js
@@ -1,6 +1,7 @@
 // Components
 import ConditionalWrap from './conditional-wrap.js';
 import SelectItemControl from './select-item-control.js';
+import TitleIndicator from './title-indicator.js';
 import ListInputControl from './list-input-control.js';
 import AdvTextControl from './advanced-text-control.js';
 import Loader from './loader.js';
@@ -11,6 +12,7 @@ import RawHTML from './raw.js';
 export {
     ConditionalWrap,
     SelectItemControl,
+    TitleIndicator,
     ListInputControl,
     AdvTextControl,
     Loader,