+++ zukit/src/scripts/components/index.js
@@ -7,6 +7,7 @@ import ZukitTable from './table.js';
 
 import ConditionalWrap from './conditional-wrap.js';
 import SelectItemControl from './select-item-control.js';
+import TitleIndicator from './title-indicator.js';
 import ListInputControl from './list-input-control.js';
 import AdvTextControl from './advanced-text-control.js';
 import Loader from './loader.js';
@@ -23,6 +24,7 @@ export {
 
     ConditionalWrap,
     SelectItemControl,
+    TitleIndicator,
     ListInputControl,
     AdvTextControl,
     Loader,