+++ zukit/src/scripts/render.js
@@ -8,16 +8,20 @@ const { ToggleControl, SelectControl, Button } = wp.components;
 // Internal dependencies
 
 import { externalData, checkDependency, simpleMarkdown } from './utils.js';
-import { setRestRouter } from './fetch.js';
+import { setRestBasics } from './fetch.js';
 import ZukitSkeleton from './components/skeleton.js'
 import ZukitDivider from './components/divider.js'
 
+export function externalDataSettings(pageId) {
+	return externalData(`${pageId}_settings`);
+}
+
 export function renderPage(pageId, settings = {}) {
 
-	const pageData = externalData(`${pageId}_settings`);
+	const pageData = externalDataSettings(pageId);
 
 	// restRouter serves to identify the plugin/theme that currently uses the REST API
-	setRestRouter(pageData.router);
+	setRestBasics(pageData);
 
 	if(get(settings, 'panels') !== undefined) {
 		// Get 'debug' options key